+++ messages/params/destination-pointcode.go
@@ -15,5 +15,5 @@ func (p *Param) DestinationPointCode() uint32 {
 	if p.Tag != DestinationPointCode {
 		return 0
 	}
-	return uint32(p.decodeUint32ValData() & 0xffffff)
+	return p.decodeUint32ValData() & 0xffffff
 }
